backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pymupdf
python-pptx
python-dotenv
httpx
beautifulsoup4
duckduckgo-search
openpyxl
notebooklm-py
